backend/requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.9.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
timm==0.9.12

# Graph Neural Networks
torch-geometric==2.4.0
torch-scatter==2.1.1
torch-sparse==0.6.18
torch-cluster==1.6.3

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
albumentations==1.3.1

# NLP
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
